pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-docx-to-markdown"
version = "1.0.0"
description = "Convert PDF and DOCX files to clean Markdown locally"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "PyMuPDF>=1.23.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "tqdm>=4.66.0",
]

[project.scripts]
pdf-docx-to-markdown = "pdf_docx_to_markdown:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
